% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrondi_entier.R
\name{arrondi_entier}
\alias{arrondi_entier}
\title{Arrondit un vecteur numérique sur une cible entière}
\usage{
arrondi_entier(x, target = NULL, na.rm = F, verbose = F)
}
\arguments{
\item{x:}{vecteur à traiter contenant uniquement des valeurs positives}

\item{target:}{entier cible pour la somme de x, par défaut arrondi de la somme de x}

\item{na.rm:}{supprime les valeurs manquantes si TRUE, FALSE par défaut}

\item{verbose:}{si TRUE, affiche des informations sur le calcul}
}
\description{
Arrondit un vecteur numérique sur une cible entière
}
