% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{as_int}
\alias{as_int}
\title{Wrapper try_cast duckdb}
\usage{
as_int(x)
}
\description{
Wrapper try_cast duckdb
}
\examples{
p90 |>
  transmute(
    NPER = !!as_int(NPER),
    NE11M = !!as_int(NE11M),
    INP75M = !!as_bool(INP75M)
  )
}
