% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{count_bits}
\alias{count_bits}
\title{Compte le nombre de bits par éléement d'un vecteur entier}
\usage{
count_bits(n)
}
\arguments{
\item{n}{Vecteur d'entiers}
}
\value{
Nombre de '1' dans l'expression binaire
}
\description{
Compte le nombre de bits par éléement d'un vecteur entier
}
\examples{
count_bits(1:16)
}
