% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_conn.R
\docType{data}
\name{db}
\alias{db}
\title{Crée une instance de duckdb_conn}
\format{
An object of class \code{duckdb_conn} (inherits from \code{R6}) of length 10.
}
\usage{
db
}
\description{
Crée une instance de duckdb_conn
}
\keyword{datasets}
