% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{db_param}
\alias{db_param}
\title{Renvoie une liste contenant les paramètres de connexion Postgres}
\usage{
db_param(source, an = NULL)
}
\arguments{
\item{source}{nom de la source RP, par exemple "omer" ou "ear"}

\item{an}{facultatif sauf pour certaines sources RP}
}
\value{
Liste de 4 éléments nommés 'dbname', 'host', 'user' et 'port'
}
\description{
Renvoie une liste contenant les paramètres de connexion Postgres
}
