% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downzip}
\alias{downzip}
\title{Télécharge et décompresse un csv zippé}
\usage{
downzip(url, keep = F)
}
\arguments{
\item{url:}{URL du fichier zip}

\item{keep:}{TRUE pour conserver le zip téléchargé, FALSE par défaut}
}
\description{
Télécharge et décompresse un csv zippé
}
