% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{ds_register}
\alias{ds_register}
\title{Traite un dataset arrow comme une table duckdb}
\usage{
ds_register(conn, ds)
}
\arguments{
\item{ds}{: dataset arrow}
}
\description{
Traite un dataset arrow comme une table duckdb
}
