% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{ear_files}
\alias{ear_files}
\title{Renvoie une liste des fichiers EAR}
\usage{
ear_files()
}
\arguments{
\item{conn}{: connexion duckdb, peut être obtenu par la fonction get_conn}
}
\value{
Liste de 4 éléments nommés 'ind', 'log', 'fam', 'liens' pour être passée
à \code{get_tbl} ou \code{get_ds} selon exemple ci-dessous
}
\description{
Renvoie une liste des fichiers EAR
}
\examples{
conn <- get_conn()
t_ear <- tbl_list(conn, ear_files(), level=1)
t_ear$ind |>
  dplyr::group_by(an) |>
  dplyr::count() |>
  dplyr::arrange(an) |>
  print()
}
