% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{ecart}
\alias{ecart}
\title{Renvoie TRUE si écart dans la norme}
\usage{
ecart(v1, v2, abs = 0, rel = 0)
}
\arguments{
\item{v1}{: valeur initiale}

\item{v2}{: valeur finale}

\item{abs}{: écart absolu maximal, 0 par défaut}

\item{rel}{: écart relatif maximal, 0 par défaut}
}
\value{
Renvoie TRUE si écart dans la norme, FALSE sinon
}
\description{
Renvoie TRUE si écart dans la norme
}
\examples{
# OK si ecart < 5\% ou <= 10
ecart(c(100, 100, 100), c(105, 115, 80), 10, 0.05)
# [1]  TRUE FALSE FALSE
}
