% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{extend}
\alias{extend}
\title{Renvoie une liste nommée à partir d'un pattern}
\usage{
extend(l, pattern)
}
\arguments{
\item{l}{: vecteur character}

\item{pattern}{: pattern glue où "{x}" est remplacé par les valeurs de l}
}
\value{
Liste d'éléments nommés avec le pattern résolu pour chauqe valeur
}
\description{
Renvoie une liste nommée à partir d'un pattern
}
\examples{
extend(c("a", "b", "c"), "test{x}")
}
