% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp21.R
\name{get_ds}
\alias{get_ds}
\title{Renvoie une liste de datasets RP EDL}
\usage{
get_ds(files)
}
\arguments{
\item{files}{: liste des fichiers obtenu par la fonction rp_edl_files(an) par
exemple}
}
\value{
Liste de 5 éléments nommés 'pind', 'plog', 'cind', 'clog', 'cfam'
selon principal (p)/complémentaire(c) et individu(ind)/logement(log)/famille(fam)
}
\description{
Renvoie une liste de datasets RP EDL
}
\examples{
ds_rp2015 <- get_ds(rp_edl_file(2015))
}
