% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{get_ear}
\alias{get_ear}
\title{Renvoit tables EAR depuis GEN}
\usage{
get_ear(conn, an, src = "gen")
}
\arguments{
\item{conn}{: connexion duckdb, peut être obtenu par la fonction get_conn}
}
\value{
Liste de 4 éléments nommés 'ind', 'log', 'fam' et 'liens'
}
\description{
Renvoit tables EAR depuis GEN
}
