% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{get_eardb}
\alias{get_eardb}
\title{Renvoie une liste des tables EAR depuis base duckdb}
\usage{
get_eardb(conn, an, db = "ear", crypt = FALSE)
}
\arguments{
\item{conn}{: connexion duckdb, peut être obtenu par la fonction get_conn}
}
\value{
Liste de 4 éléments nommés 'ind', 'log', 'fam' et 'liens'
}
\description{
Renvoie une liste des tables EAR depuis base duckdb
}
