% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{get_rp}
\alias{get_rp}
\title{Renvoie une liste des tables du RP21}
\usage{
get_rp(conn, an = 2021, src = c("gen", "edl"))
}
\arguments{
\item{conn}{: connexion duckdb, peut être obtenu par la fonction get_conn}
}
\value{
Liste de 5 éléments nommés 'pind', 'plog', 'cind', 'clog', 'cfam'
selon principal (p)/complémentaire(c) et individu(ind)/logement(log)/famille(fam)
}
\description{
Renvoie une liste des tables du RP21
}
\note{
La variable d'environnement SITE doit être défini à 'ls3' ou 'aus'.
Par défaut, 'aus'.
}
\examples{
conn <- get_conn()
trp21 <- get_rp21(conn)
# population France
trp21$pind |>
  dplyr::count(wt = ipondi)
}
