% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp21.R
\name{get_rp21}
\alias{get_rp21}
\title{Renvoie une liste des tables du RP21}
\usage{
get_rp21(conn)
}
\arguments{
\item{conn}{: connexion duckdb, peut être obtenu par la fonction get_conn}
}
\description{
Renvoie une liste des tables du RP21
}
\note{
La variable d'environnement SITE doit être défini à 'ls3' ou 'aus'.
Par défaut, 'aus'.
}
