% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp21.R
\name{get_tbl}
\alias{get_tbl}
\title{Renvoie une liste des tables duckdb du coffre RP EDL}
\usage{
get_tbl(conn, files, level = 1)
}
\arguments{
\item{conn}{: connexion duckdb, peut être obtenu par la fonction get_conn}

\item{files}{: liste des fichiers obtenu par la fonction rp_edl_files(an) par}

\item{level}{: nombre de niveaux dans le cas de fichiers parquet partitionnés,
par défaut 1 si pas de partionnement}
}
\value{
Liste de 5 éléments nommés 'pind', 'plog', 'cind', 'clog', 'cfam'
selon principal (p)/complémentaire(c) et individu(ind)/logement(log)/famille(fam)
}
\description{
Renvoie une liste des tables duckdb du coffre RP EDL
}
\examples{
conn <- get_conn()
trp2015 <- get_tbl(conn, rp_edl_files(2015))
trp2015$pind |>
  dplyr::count(wt=ipondi) |>
  dplyr::collect()
}
