% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{group_num}
\alias{group_num}
\title{Crée une valeur de 'grouping' par facililter l'utilsation des fonctions 'cube'}
\usage{
group_num(x)
}
\arguments{
\item{x:}{Vecteur de booleans avec TRUE pour les dimensions présentes}
}
\value{
Valeur de groupe correspondant
}
\description{
Crée une valeur de 'grouping' par facililter l'utilsation des fonctions 'cube'
}
\examples{
# Croisement région x statut occupation
group_num(c(reg=T, dep=F, densaav=F, maisappt=F, achev=F, typc=F, stoc=T, surf=F))
# 125 = 255 - (128 + 2)
}
