% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{install_extensions}
\alias{install_extensions}
\title{Installe extensions duckdb depuis le répertoire ~/extensions}
\usage{
install_extensions()
}
\description{
Installe extensions duckdb depuis le répertoire ~/extensions
}
