% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{lapplyn}
\alias{lapplyn}
\title{Fonction lapply avec utilisation des noms de la liste nommée}
\usage{
lapplyn(l, f)
}
\arguments{
\item{l}{liste character}

\item{f}{fonction appliquée à chaque élément de la liste}
}
\value{
Liste d'éléments nommés avec le résultat de la fonction
}
\description{
Fonction lapply avec utilisation des noms de la liste nommée
}
