% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\docType{data}
\name{list_db}
\alias{list_db}
\title{Liste des sources de données RP simples (sans paramètre an)}
\format{
An object of class \code{list} of length 20.
}
\usage{
list_db
}
\description{
Liste des sources de données RP simples (sans paramètre an)
}
\keyword{datasets}
