% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{pg_attach}
\alias{pg_attach}
\title{Attache une base Postgres à duckdb à partir dune liste créée par \code{\link[=db_param]{db_param()}}}
\usage{
pg_attach(conn, param, db = "pg", schema = "public")
}
\arguments{
\item{conn:}{connexion duckdb, peut être obtenu par la fonction get_conn}

\item{param:}{liste de paramètres créée par \code{\link[=db_param]{db_param()}}}

\item{db:}{nom de l'alias de la base}

\item{schema:}{nom du schema Postgres, 'public' par défaut}
}
\value{
Code retour dbExecute
}
\description{
Attache une base Postgres à duckdb à partir dune liste créée par \code{\link[=db_param]{db_param()}}
}
\examples{
conn = DBI::dbConnect(duckdb::duckdb())
pg_attach(conn, db_param("ear", 2025), db="ear", schema = "rp")
pg_attach(conn, db_param("rorcal_clone"), db="rorcal")
}
