% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{pg_connect}
\alias{pg_connect}
\title{Renvoie une connexion Postgres à partir dune liste créée par \code{\link[=db_param]{db_param()}}}
\usage{
pg_connect(param)
}
\arguments{
\item{param}{liste de paramètres créée par \code{\link[=db_param]{db_param()}}}
}
\value{
Connexion Postgres
}
\description{
Renvoie une connexion Postgres à partir dune liste créée par \code{\link[=db_param]{db_param()}}
}
\examples{
conn_pg <- pg_connect(db_param("rorcal_clone"))
conn_ear2025 <- pg_connect(db_param("ear", an=2025))
}
