% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_st.R
\name{read_s3_st}
\alias{read_s3_st}
\title{Lit une table de classe sf depuis stockage s3}
\usage{
read_s3_st(name, s3path = "gpkg")
}
\arguments{
\item{name}{: nom du fichier à lire}

\item{s3path}{: chemin du stockage s3, par défaut \code{gpkg}}
}
\description{
Lit une table de classe sf depuis stockage s3
}
