% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp_edl_files}
\alias{rp_edl_files}
\title{Renvoie une liste des noms de fichiers parquet du coffre RP EDL}
\usage{
rp_edl_files(an)
}
\arguments{
\item{an}{: année du RP, exemple 2015}
}
\value{
Liste de 5 éléments nommés 'pind', 'plog', 'cind', 'clog', 'cfam'
selon principal (p)/complémentaire(c) et individu(ind)/logement(log)/famille(fam)
}
\description{
Renvoie une liste des noms de fichiers parquet du coffre RP EDL
}
\examples{
conn <- get_conn()
trp2015 <- get_tbl(conn, rp_edl_files(2015))
}
