% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3download}
\alias{s3download}
\title{Télécharge un fichier dans s3 si absent}
\usage{
s3download(url, path, fs, force = FALSE)
}
\arguments{
\item{url}{: url à télécharger}

\item{path}{: chemin s3 à créer, si absent prend le dernier nom de l'url}

\item{fs}{: filesystem, par défaut s3fs}

\item{force}{: force le téléchargement même si fichier déjà présent}
}
\value{
Renvoit un SubTreeFileSystem utilisable avec \code{copy_files}
}
\description{
Télécharge un fichier dans s3 si absent
}
