% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3file}
\alias{s3file}
\title{Renvoit un SubFileSystemTree}
\usage{
s3file(path, root, fs)
}
\arguments{
\item{path}{chemin}
}
\description{
Renvoit un SubFileSystemTree
}
\note{
Peut être utilisé avec read_parquet, write_parquet
et read_csv_arrow
}
