% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\docType{data}
\name{s3fs}
\alias{s3fs}
\title{Crée un filesystem s3 avec les variables environnement}
\format{
An object of class \code{S3FileSystem} (inherits from \code{FileSystem}, \code{ArrowObject}, \code{R6}) of length 27.
}
\usage{
s3fs
}
\description{
Crée un filesystem s3 avec les variables environnement
}
\keyword{datasets}
