% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3path}
\alias{s3path}
\title{Renvoit une chaine s3 depuis un chemin}
\usage{
s3path(path, bucket, prefix)
}
\arguments{
\item{path}{: chemin}
}
\value{
chaîne chemin
}
\description{
Renvoit une chaine s3 depuis un chemin
}
