% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{set_conn}
\alias{set_conn}
\title{Crée une connexion globale duckdb 'conn'}
\usage{
set_conn(ext = c("core", "all", "none"))
}
\arguments{
\item{ext:}{indique les extensions chargées. 'core' = httpfs et spatial,
'all' = ajoute h3 et read_stat, 'none' = pas d'extension. 'core' par défaut}
}
\description{
Crée une connexion globale duckdb 'conn'
}
