% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_db}
\alias{tbl_db}
\title{Renvoie une liste de tables depuis une database duckdb}
\usage{
tbl_db(conn, db, verbose = FALSE)
}
\arguments{
\item{conn}{: connexion duckdb, peut être obtenu par la fonction get_conn}

\item{db:}{nom de la database}
}
\value{
Liste de tables dplyr
}
\description{
Renvoie une liste de tables depuis une database duckdb
}
