% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_duckdb}
\alias{tbl_duckdb}
\title{Renvoie une liste de tables depuis une base duckdb}
\usage{
tbl_duckdb(conn, name, db = NULL, crypt = FALSE)
}
\arguments{
\item{conn}{: connexion duckdb, peut être obtenu par la fonction get_conn}
}
\value{
Liste de tables
}
\description{
Renvoie une liste de tables depuis une base duckdb
}
