% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_expl}
\alias{tbl_expl}
\title{Renvoit une table duckdb depuis S3 exploitations stat}
\usage{
tbl_expl(conn, path)
}
\arguments{
\item{conn}{: connexion duckdb}

\item{path}{: chemin de la table/répertoire parquet}
}
\description{
Renvoit une table duckdb depuis S3 exploitations stat
}
