% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_pqt}
\alias{tbl_pqt}
\title{Renvoit une table duckdb depuis un fichier parquet y.c S3}
\usage{
tbl_pqt(conn, path, lowercase = TRUE)
}
\arguments{
\item{conn}{: connexion duckdb}

\item{path}{: chemin de la table/répertoire parquet}

\item{tolower}{: si TRUE, le défaut, les variables sont converties
en minuscules}
}
\description{
Renvoit une table duckdb depuis un fichier parquet y.c S3
}
