% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_s3}
\alias{tbl_s3}
\title{Renvoit une table duckdb depuis S3 personnel}
\usage{
tbl_s3(conn, path)
}
\arguments{
\item{conn}{: connexion duckdb}

\item{path}{: chemin de la table/répertoire parquet}
}
\description{
Renvoit une table duckdb depuis S3 personnel
}
