% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{tempname}
\alias{tempname}
\title{Renvoit un nom temporaire}
\usage{
tempname(n = 8, prefix = "temp")
}
\arguments{
\item{n}{Longueur du nom, par défaut 8}

\item{prefix}{Préfixe éventuel, par défaut 'temp'}
}
\value{
Chaîne aléatoire de longueur n préfixée par prefix
}
\description{
Renvoit un nom temporaire
}
\examples{
tempname(4, "cube")
[1] "cube_poiu"
}
