% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{txevol}
\alias{txevol}
\title{Calcul taux évolution}
\usage{
txevol(eff1, eff2, nb_per, pct = F, dec = NA)
}
\arguments{
\item{eff1:}{effectif initial}

\item{eff2:}{effectif final}

\item{nb_per:}{nombre de périodes}

\item{pct:}{évolution en \%, FALSE par défaut}

\item{dec:}{nombre de décimales, par défaut toutes}
}
\description{
Calcul taux évolution
}
