% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparquet.R
\name{write_geoparquet}
\alias{write_geoparquet}
\title{Permet d'écrire un fichier geoparquet en sf}
\usage{
write_geoparquet(sf, path, verbose = F)
}
\description{
Permet d'écrire un fichier geoparquet en sf
}
