% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_st.R
\name{write_s3_st}
\alias{write_s3_st}
\title{Écrit une table de classe sf vers stockage s3}
\usage{
write_s3_st(st_table, name, s3path = "gpkg")
}
\arguments{
\item{st_table}{table de class st}

\item{name}{nom du fichier à écrire en format gpkg}

\item{s3path}{chemin du stockage s3, par défaut \code{gpkg}}
}
\description{
Écrit une table de classe sf vers stockage s3
}
