% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrondi_entier.R
\name{arrondi_entier}
\alias{arrondi_entier}
\title{Arrondit un vecteur numérique sur une cible entière}
\usage{
arrondi_entier(
  x,
  p = NULL,
  method = c("abs", "max", "min"),
  target = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{x}{vecteur à traiter contenant uniquement des valeurs positives}

\item{p}{(optionnel) vecteur de poids, sinon poids = 1}

\item{target}{(optionnel) entier cible pour la somme pondérée de x,
par défaut somme pondérée de x}

\item{na.rm}{(optionnel) supprime les valeurs manquantes si TRUE, FALSE par défaut}
}
\description{
Arrondit un vecteur numérique sur une cible entière
}
