% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{attach_db}
\alias{attach_db}
\title{Attach une base duckdb à une connexion}
\usage{
attach_db(conn, name, db = NULL, crypt = FALSE)
}
\arguments{
\item{conn:}{connexion duckdb, peut être obtenu par la fonction get_conn}

\item{path:}{nom de la base duckdb, ".duckdb" est ajouté automatiquement et la base est dans le répertoire
(s3perso)/duckdb}

\item{db:}{nom de l'alias de la base, par défaut path}

\item{crypt:}{si TRUE, passe la valeur de DUCKDB_ENCRYPTION_KEY comme clé de la base cryptée}
}
\value{
Nom de la base duckdb
}
\description{
Attach une base duckdb à une connexion
}
