% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{download}
\alias{download}
\title{Télécharge un fichier si absent}
\usage{
download(url, file, dir = tempdir(), force = FALSE, verbose = FALSE)
}
\arguments{
\item{url}{URL du fichier}

\item{file}{Chemin du fichier destination, par défaut extrait de l'url}

\item{dir}{Chemin du fichier destination, par défaut répertoire courant}

\item{force}{TRUE pour forcer le téléchargement même si le fichier existe, FALSE par défaut}

\item{verbose}{TRUE pour afficher davantage d'information, FALSE par défaut}
}
\description{
Télécharge un fichier si absent
}
