% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{download_archive}
\alias{download_archive}
\title{Télécharge et décompresse une fichier compressé}
\usage{
download_archive(url, dir = NULL, force = FALSE, verbose = FALSE)
}
\arguments{
\item{url:}{URL du fichier}

\item{keep:}{TRUE pour conserver le fichier téléchargé et l'extraction}
}
\description{
Télécharge et décompresse une fichier compressé
}
