% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp21.R
\name{ear_files}
\alias{ear_files}
\title{Renvoie une liste des fichiers EAR}
\usage{
ear_files()
}
\arguments{
\item{conn}{: connexion duckdb, peut être obtenu par la fonction get_conn}
}
\value{
Liste de 4 éléments nommés 'ind', 'log', 'fam', 'liens' pour être passée
à \code{get_tbl} ou \code{get_ds} selon exemple ci-dessous
}
\description{
Renvoie une liste des fichiers EAR
}
\note{
La variable d'environnement SITE doit être défini à 'ls3' ou 'aus'.
Par défaut, 'aus'.
}
\examples{
conn <- get_conn()
tear <- get_tbl(conn, ear_files(), level=2)
tear$ind |>
  dplyr::group_by(an) |>
  dplyr::count() |>
  dplyr::arrange(an) |>
  print()

s3fs <- s3createfs()
dsear = lapply(ear_files(), function(x) arrow::open_dataset(s3fs$path(substr(x, 6, nchar(x)))))
dsear$ind |>
  dplyr::group_by(an) |>
  dplyr::count() |>
  dplyr::arrange(an) |>
  dplyr::collect() |>
  print()
}
