% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{get_conn}
\alias{get_conn}
\title{Renvoie une connexion duckdb}
\usage{
get_conn(dbdir = ":memory:", ext = c("none", "core", "all"))
}
\arguments{
\item{dbdir}{: nom du la base duckdb, par défaut base stockée en mémoire}

\item{ext:}{indique les extensions chargées. 'core' = httpfs et spatial,
'all' = ajoute h3 et read_stat, 'none' = pas d'extension. 'none' par défaut}
}
\description{
Renvoie une connexion duckdb
}
