% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp21.R
\name{get_conn}
\alias{get_conn}
\title{Renvoie une connexion duckdb}
\usage{
get_conn(dbdir = ":memory:")
}
\arguments{
\item{dbdir}{: nom du la base duckdb, par défaut base stockée en mémoire}
}
\description{
Renvoie une connexion duckdb
}
