% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{get_eardb}
\alias{get_eardb}
\title{Renvoie une liste des tables RP depuis base duckdb}
\usage{
get_eardb(conn, an, db = "ear")
}
\arguments{
\item{conn}{: connexion duckdb, peut être obtenu par la fonction get_conn}
}
\value{
Liste de 5 éléments nommés 'pind', 'plog', 'cind', 'clog', 'cfam'
selon principal (p)/complémentaire(c) et individu(ind)/logement(log)/famille(fam)
}
\description{
Renvoie une liste des tables RP depuis base duckdb
}
