% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparquet.R
\name{read_geoparquet}
\alias{read_geoparquet}
\title{Permet de lire un fichier geoparquet en sf}
\usage{
read_geoparquet(
  path,
  crs = NULL,
  transform = F,
  xy = F,
  as.tibble = T,
  verbose = F
)
}
\description{
Permet de lire un fichier geoparquet en sf
}
