% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3file}
\alias{s3file}
\title{Copie locale d'un chemin s3}
\usage{
s3file(path, fs, ...)
}
\arguments{
\item{path}{: chemin}
}
\description{
Copie locale d'un chemin s3
}
\note{
Peut être utilisé avec read_parquet, write_parquet
      et read_csv_arrow
}
