% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{set_conn}
\alias{set_conn}
\title{Crée une connexion globale duckdb 'conn'}
\usage{
set_conn(ext = c("none", "core", "geo"))
}
\arguments{
\item{ext:}{indique les extensions chargées. 'core': spatial, 'geo': h3 et spatial,
'none' = pas d'extension. 'none' par défaut}
}
\description{
Crée une connexion globale duckdb 'conn'
}
