% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smb.R
\name{smb}
\alias{smb}
\title{Commandes vers partages Windows}
\usage{
smb(cmd, path, dir = ".")
}
\arguments{
\item{cmd}{Commande parmi "get" "mget" "put" "mput" "mkdir" "ls" "lsvol"}

\item{path}{Chemin de fichier}

\item{dir}{Répertoire}
}
\description{
Commandes vers partages Windows
}
\details{
lsvol liste la racine d'un lecteur avec la syntaxe smb("lsvol", "/X/*/")
}
