% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_duckdb_parquet.R
\name{write_duckdb_parquet}
\alias{write_duckdb_parquet}
\title{Écrit fichier parquet depuis une table duckdb}
\usage{
write_duckdb_parquet(
  conn,
  name = NULL,
  filename = NULL,
  dir = ".",
  partition = NULL,
  verbose = FALSE
)
}
\arguments{
\item{conn}{: connexion duckdb}

\item{name}{: nom de la table}

\item{filename}{: nom du fichier ou répertoire si partition est renseigné, par défaut \code{name}}

\item{dir}{: répertoire de sortie, éventuellement s3}

\item{partition}{: chaine de la forme "var1, var2" pour créer une partition, pas de partition par défaut}

\item{verbose}{: indique la requete SQL générée}
}
\value{
filename
}
\description{
Écrit fichier parquet depuis une table duckdb
}
