% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_duckdb_parquet.R
\name{write_duckdb_parquet}
\alias{write_duckdb_parquet}
\title{Écrit fichier parquet depuis une table duckdb}
\usage{
write_duckdb_parquet(
  table,
  path = NULL,
  dir = ".",
  partition = NULL,
  order_by = NULL,
  verbose = FALSE
)
}
\arguments{
\item{table}{: table duckdb}

\item{path}{: nom du fichier ou répertoire si partition est renseigné, par défaut le nom de la table}

\item{dir}{: répertoire de sortie, éventuellement s3}

\item{partition}{: chaine de la forme "var1, var2" pour créer une partition, pas de partition par défaut}

\item{order_by}{: ordre éventuel de sortie sous forme de vecteur character}

\item{verbose}{: indique la requete SQL générée}
}
\value{
nom du fichier/répertoire
}
\description{
Écrit fichier parquet depuis une table duckdb
}
\examples{
s3expl = "s3://insee/sern-div-exploitations-statistiques-rp/ear_rp"
# Fichier simple ear_fam.parquet dasn répertoire dir sous s3
outfile = write_duckdb_parquet(ear_fam, dir = s3expl)
# Fichier simple ear_fam.parquet trié dans répertoire courant
outfile = write_duckdb_parquet(ear_fam, order_by=c("var1", "var2"))
# Fichier partitionné dans répertoire test_fam
outfile = write_duckdb_parquet(ear_fam,  dir = s3expl,
  filename="test_fam", partition="c_annee_col", verbose=T)
}
